requirements.txt
aniso8601==9.0.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
axios==0.3.0
backcall==0.2.0
beautifulsoup4==4.12.0
binaryornot==0.4.4
bleach==6.0.0
blis==0.7.9
CacheControl==0.14.0
cachetools==5.3.3
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cookiecutter==2.1.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
dash==2.8.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateparser==1.1.8
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
et_xmlfile==2.0.0
executing==1.2.0
fastjsonschema==2.16.3
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.9.0
firebase-admin==6.5.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.39.3
fqdn==1.5.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
GoogleNews==1.6.7
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.13.3
idna==3.4
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
jieba3k==0.35.1
Jinja2==3.1.2
jinja2-time==0.2.0
jiter==0.8.2
joblib==1.2.0
jsonpickle==3.0.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
langcodes==3.3.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.9
nbclassic==0.5.4
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
newspaper3k==0.2.8
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.2
openai==1.61.1
openpyxl==3.1.5
orjson==3.8.10
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.13.0
plyer==2.1.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
PyJWT==2.8.0
pyparsing==3.0.9
pyrsistent==0.19.3
pystray==0.19.5
python-dateutil==2.8.2
python-docx==1.1.2
python-json-logger==2.0.7
python-slugify==8.0.0
pytz==2023.2
pytz-deprecation-shim==0.1.0.post0
pyvis==0.3.2
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.3
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
Send2Trash==1.8.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.1
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.2
srsly==2.4.6
stack-data==0.6.2
sympy==1.11.1
tenacity==8.2.1
terminado==0.17.1
text-unidecode==1.3
thinc==8.1.9
threadpoolctl==3.4.0
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==3.4.0
tokenizers==0.13.2
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
typer==0.7.0
typing_extensions==4.12.2
tzdata==2023.2
tzlocal==4.3
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.14
vboxapi==1.0
virtualenv==20.19.0
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.7
wikipedia==1.4.0
